testqwen/streamlit_qwen2_withbyaldi/requirements.txt
streamlit
torch
torchvision
torchaudio
torchao
git+https://github.com/huggingface/transformers.git
diffusers
Pillow
qwen_vl_utils
nltk
byaldi